# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CityCard API",
        "version": "1.0"
    },
    "paths": {
...
        "/admin/cities/store": {
            "post": {
                "tags": [
                    "Cities"
                ],
                "summary": "Store a new city",
                "operationId": "storeCity",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Name of the city",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "City created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/City"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/cities/update/{city}": {
            "patch": {
                "tags": [
                    "Cities"
                ],
                "summary": "Update an existing city",
                "operationId": "updateCity",
                "parameters": [
                    {
                        "name": "city",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Name of the city",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "City updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/City"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/cities/destroy/{city}": {
            "delete": {
                "tags": [
                    "Cities"
                ],
                "summary": "Delete an existing city",
                "operationId": "deleteCity",
                "parameters": [
                    {
                        "name": "city",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "City deleted successfully"
                    }
                }
            }
        },
...
        "/admin/tickets/store": {
            "post": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Store a new ticket",
                "operationId": "storeTicket",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "card_id"
                                ],
                                "properties": {
                                    "type": {
                                        "description": "Ticket type",
                                        "type": "string"
                                    },
                                    "card_id": {
                                        "description": "ID of the card associated with the ticket",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ticket created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ticket"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/tickets/update/{ticket}": {
            "patch": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Update an existing ticket",
                "operationId": "updateTicket",
                "parameters": [
                    {
                        "name": "ticket",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type"
                                ],
                                "properties": {
                                    "type": {
                                        "description": "Ticket type",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ticket updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ticket"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/cities/destroy/{ticket}": {
            "delete": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Delete an existing ticket",
                "operationId": "deleteTicket",
                "parameters": [
                    {
                        "name": "ticket",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "City deleted successfully"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "City": {
                "title": "City",
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "City ID",
                        "type": "integer"
                    },
                    "name": {
                        "description": "City name",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Timestamp when the city was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Timestamp when the city was updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Ticket": {
                "title": "Ticket",
                "required": [
                    "id",
                    "card_id",
                    "type"
                ],
                "properties": {
                    "id": {
                        "description": "Ticket ID",
                        "type": "integer"
                    },
                    "card_id": {
                        "description": "ID of the card associated with the ticket",
                        "type": "integer"
                    },
                    "type": {
                        "description": "Ticket type",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Timestamp when the ticket was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Timestamp when the ticket was updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}